requirements.txt
moderngl
pyglet
pypng
importlib_resources; python_version < '3.9'
